+++ packages/core-js/actual/promise/index.js
@@ -1,5 +1,6 @@
 'use strict';
 var parent = require('../../stable/promise');
+require('../../modules/esnext.promise.try');
 require('../../modules/esnext.promise.with-resolvers');
 
 module.exports = parent;