+++ packages/core-js/full/promise/index.js
@@ -3,7 +3,6 @@ var parent = require('../../actual/promise');
 // TODO: Remove from `core-js@4`
 require('../../modules/esnext.aggregate-error');
 require('../../modules/esnext.promise.all-settled');
-require('../../modules/esnext.promise.try');
 require('../../modules/esnext.promise.any');
 
 module.exports = parent;