+++ packages/core-js/full/promise/try.js
@@ -1,13 +1,4 @@
 'use strict';
-require('../../modules/es.promise');
-require('../../modules/esnext.promise.try');
-var call = require('../../internals/function-call');
-var isCallable = require('../../internals/is-callable');
-var path = require('../../internals/path');
+var parent = require('../../actual/promise/try');
 
-var Promise = path.Promise;
-var promiseTry = Promise['try'];
-
-module.exports = { 'try': function (callbackfn) {
-  return call(promiseTry, isCallable(this) ? this : Promise, callbackfn);
-} }['try'];
+module.exports = parent;