+++ packages/core-js/stage/2.7.js
@@ -2,6 +2,5 @@
 var parent = require('./3');
 
 require('../proposals/math-sum');
-require('../proposals/promise-try');
 
 module.exports = parent;