+++ packages/core-js/stage/3.js
@@ -8,6 +8,7 @@ require('../proposals/explicit-resource-management');
 require('../proposals/float16');
 require('../proposals/iterator-helpers-stage-3-2');
 require('../proposals/json-parse-with-source');
+require('../proposals/promise-try');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');
 require('../proposals/array-grouping-stage-3-2');